Fix wrong method name
The docs says *ShallowDeepAlmostEqual* but the code says *shallowDeepAlmostEqual*. The latter is in line with the original [shallowDeepEqual](https://github.com/michelsalib/chai-shallow-deep-equal). My tests failed because of this.

P.S. thanks for the tool! By the way, you should enable issues from the repo settings to let people to report bugs easily.